Refactoring and API surfacing
